requirements/dev.txt
Django==3.2.11
djangorestframework==3.13.1
epub-meta==0.0.7
grequests==0.6.0
PyPDF2==1.26.0
